Revert "HBASE-13661 Correct binary compatibility issues discovered in 1.1.0RC0"
This reverts commit 5c7e57240cf494200ec31677801ded13cf2e9201.

+++ hbase-client/src/main/java/org/apache/hadoop/hbase/ServerLoad.java
@@ -86,7 +86,7 @@ public class ServerLoad {
   protected ClusterStatusProtos.ServerLoad serverLoad;
 
   /* @return number of requests  since last report. */
-  public long getNumberOfRequests() {
+  public int getNumberOfRequests() {
     return serverLoad.getNumberOfRequests();
   }
   public boolean hasNumberOfRequests() {
@@ -94,7 +94,7 @@ public class ServerLoad {
   }
 
   /* @return total Number of requests from the start of the region server. */
-  public long getTotalNumberOfRequests() {
+  public int getTotalNumberOfRequests() {
     return serverLoad.getTotalNumberOfRequests();
   }
   public boolean hasTotalNumberOfRequests() {
