+++ hbase-hadoop-compat/src/main/java/org/apache/hadoop/hbase/master/MetricsMasterSource.java
@@ -79,9 +79,9 @@ public interface MetricsMasterSource extends BaseSource {
    *
    * @param inc Ammount to increment the total by.
    */
-  void incRequests(final long inc);
+  void incRequests(final int inc);
 
-
+  
 
 
 }
